requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
branca==0.7.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
chatbotAI==0.3.1.3
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateutils==0.6.12
decorator==5.1.1
diffusers==0.30.0
distro==1.9.0
duckdb==1.0.0
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
folium==0.17.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.65.5
gw_dsl_parser==0.1.49.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.3
ipylab==1.0.0
ipython==8.26.0
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.11
kanaries_track==0.0.5
keras==2.14.0
kiwisolver==1.4.5
langchain==0.3.0
langchain-core==0.3.1
langchain-text-splitters==0.3.0
langsmith==0.1.122
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
monotonic==1.6
mss==9.0.1
multidict==6.0.5
namex==0.0.8
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.46.0
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pillow==10.4.0
plotly==5.23.0
plotly-express==0.4.1
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pygwalker==0.4.9.5
pyogrio==0.9.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyproj==3.6.1
pyscreenshot==3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
quickjs==1.19.4
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
segment-analytics-python==2.2.3
shapely==2.0.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
sqlglot==25.13.0
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.37.0
streamlit-lottie==0.0.5
tenacity==8.5.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasmtime==23.0.0
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wikipedia==1.4.0
wrapt==1.14.1
xlrd==2.0.1
xyzservices==2024.6.0
yarl==1.9.4
zipp==3.19.2
